#485 move agg_insert queries to tran_sqlstr.str
