+++ src/f10_etl/tran_sqlstrs.py
@@ -350,3 +350,47 @@ def get_set_inconsistency_error_message_sqlstrs() -> dict[str, str]:
         "fiscal_timeline_month": FISCALMONT_SET_INCONSISTENCY_ERROR_MESSAGE_SQLSTR,
         "fiscal_timeline_weekday": FISCALWEEK_SET_INCONSISTENCY_ERROR_MESSAGE_SQLSTR,
     }
+
+
+FISCALUNIT_AGG_INSERT_SQLSTR = """INSERT INTO fiscalunit_agg (fiscal_title, fund_coin, penny, respect_bit, present_time, bridge, c400_number, yr1_jan1_offset, monthday_distortion, timeline_title)
+SELECT fiscal_title, MAX(fund_coin), MAX(penny), MAX(respect_bit), MAX(present_time), MAX(bridge), MAX(c400_number), MAX(yr1_jan1_offset), MAX(monthday_distortion), MAX(timeline_title)
+FROM fiscalunit_staging
+WHERE error_message IS NULL
+GROUP BY fiscal_title
+;
+"""
+FISCALDEAL_AGG_INSERT_SQLSTR = """INSERT INTO fiscal_deal_episode_agg (fiscal_title, owner_name, time_int, quota)
+SELECT fiscal_title, owner_name, time_int, MAX(quota)
+FROM fiscal_deal_episode_staging
+WHERE error_message IS NULL
+GROUP BY fiscal_title, owner_name, time_int
+;
+"""
+FISCALCASH_AGG_INSERT_SQLSTR = """INSERT INTO fiscal_cashbook_agg (fiscal_title, owner_name, acct_name, time_int, amount)
+SELECT fiscal_title, owner_name, acct_name, time_int, MAX(amount)
+FROM fiscal_cashbook_staging
+WHERE error_message IS NULL
+GROUP BY fiscal_title, owner_name, acct_name, time_int
+;
+"""
+FISCALHOUR_AGG_INSERT_SQLSTR = """INSERT INTO fiscal_timeline_hour_agg (fiscal_title, hour_title, cumlative_minute)
+SELECT fiscal_title, hour_title, MAX(cumlative_minute)
+FROM fiscal_timeline_hour_staging
+WHERE error_message IS NULL
+GROUP BY fiscal_title, hour_title
+;
+"""
+FISCALMONT_AGG_INSERT_SQLSTR = """INSERT INTO fiscal_timeline_month_agg (fiscal_title, month_title, cumlative_day)
+SELECT fiscal_title, month_title, MAX(cumlative_day)
+FROM fiscal_timeline_month_staging
+WHERE error_message IS NULL
+GROUP BY fiscal_title, month_title
+;
+"""
+FISCALWEEK_AGG_INSERT_SQLSTR = """INSERT INTO fiscal_timeline_weekday_agg (fiscal_title, weekday_title, weekday_order)
+SELECT fiscal_title, weekday_title, MAX(weekday_order)
+FROM fiscal_timeline_weekday_staging
+WHERE error_message IS NULL
+GROUP BY fiscal_title, weekday_title
+;
+"""
